+++ MK4duo/Configuration_Feature.h
@@ -34,8 +34,9 @@
  * - Solenoid extruder
  * - Color Mixing Extruder
  * - Multiextruder old MKR4
- * - Multiextruder new MKR6
- * - Multiextruder new MKSE6 (multiextruder with Servo)
+ * - Multiextruder MKR6
+ * - Multiextruder MKR12
+ * - Multiextruder MKSE6 (multiextruder with Servo)
  * - Multiextruder NPr2
  * - Multiextruder DONDOLO
  * - Extruder idle oozing prevention
@@ -270,7 +271,7 @@
  ***********************************************************************
  *                                                                     *
  * Setting for more extruder width relay system                        *
- * This is new system for 6 extruder width 2 driver and 6 relay.       *
+ * This is new system for 6 extruder width 2 driver and 8 relay.       *
  * See Configuration_pins.h for pin command relay                      *
  *                                                                     *
  * Uncomment MKR6 to enable this feature                               *
@@ -278,6 +279,23 @@
  * Uncomment INVERTED_RELE_PINS if your relay switches with GND        *
  ***********************************************************************/
 //#define MKR6
+//#define INVERTED_RELE_PINS
+/***********************************************************************/
+
+
+/***********************************************************************
+ ************************* Multiextruder MKR12 *************************
+ ***********************************************************************
+ *                                                                     *
+ * Setting for more extruder width relay system                        *
+ * This is new system for 12 extruder width 4 driver and 16 relay.     *
+ * See Configuration_pins.h for pin command relay                      *
+ *                                                                     *
+ * Uncomment MKR12 to enable this feature                              *
+ *                                                                     *
+ * Uncomment INVERTED_RELE_PINS if your relay switches with GND        *
+ ***********************************************************************/
+//#define MKR12
 //#define INVERTED_RELE_PINS
 /***********************************************************************/