+++ MK4duo/src/mechanics/core_mechanics.cpp
@@ -63,7 +63,7 @@
       if (prepare_move_to_destination_dualx()) return;
 
     #else
-      
+
       #if ENABLED(LASERBEAM) && ENABLED(LASER_FIRE_E)
         if (current_position[E_AXIS] != destination[E_AXIS] && ((current_position[X_AXIS] != destination [X_AXIS]) || (current_position[Y_AXIS] != destination [Y_AXIS]))){
           laser.status = LASER_ON;
@@ -98,6 +98,86 @@
 
     set_current_to_destination();
   }
+
+  #if ENABLED(DUAL_X_CARRIAGE)
+
+    /**
+     * Prepare a linear move in a dual X axis setup
+     */
+    bool Core_Mechanics::prepare_move_to_destination_dualx() {
+      if (active_hotend_parked) {
+        switch (dual_x_carriage_mode) {
+          case DXC_FULL_CONTROL_MODE:
+            break;
+          case DXC_AUTO_PARK_MODE:
+            if (current_position[E_AXIS] == destination[E_AXIS]) {
+              // This is a travel move (with no extrusion)
+              // Skip it, but keep track of the current position
+              // (so it can be used as the start of the next non-travel move)
+              if (delayed_move_time != 0xFFFFFFFFUL) {
+                set_current_to_destination();
+                NOLESS(raised_parked_position[Z_AXIS], destination[Z_AXIS]);
+                delayed_move_time = millis();
+                return true;
+              }
+            }
+            // unpark extruder: 1) raise, 2) move into starting XY position, 3) lower
+            for (uint8_t i = 0; i < 3; i++)
+              planner.buffer_line(
+                i == 0 ? raised_parked_position[X_AXIS] : current_position[X_AXIS],
+                i == 0 ? raised_parked_position[Y_AXIS] : current_position[Y_AXIS],
+                i == 2 ? current_position[Z_AXIS] : raised_parked_position[Z_AXIS],
+                current_position[E_AXIS],
+                i == 1 ? PLANNER_XY_FEEDRATE() : planner.max_feedrate_mm_s[Z_AXIS],
+                active_extruder,
+                active_driver
+              );
+            delayed_move_time = 0;
+            active_extruder_parked = false;
+            #if ENABLED(DEBUG_LEVELING_FEATURE)
+              if (DEBUGGING(LEVELING)) SERIAL_EM("Clear active_extruder_parked");
+            #endif
+            break;
+          case DXC_DUPLICATION_MODE:
+            if (active_extruder == 0) {
+              #if ENABLED(DEBUG_LEVELING_FEATURE)
+                if (DEBUGGING(LEVELING)) {
+                  SERIAL_MV("Set planner X", LOGICAL_X_POSITION(inactive_extruder_x_pos));
+                  SERIAL_EMV(" ... Line to X", current_position[X_AXIS] + duplicate_extruder_x_offset);
+                }
+              #endif
+              // move duplicate extruder into correct duplication position.
+              planner.set_position_mm(
+                LOGICAL_X_POSITION(inactive_extruder_x_pos),
+                current_position[Y_AXIS],
+                current_position[Z_AXIS],
+                current_position[E_AXIS]
+              );
+              planner.buffer_line(
+                current_position[X_AXIS] + duplicate_extruder_x_offset,
+                current_position[Y_AXIS], current_position[Z_AXIS], current_position[E_AXIS],
+                planner.max_feedrate_mm_s[X_AXIS], 1, active_driver
+              );
+              sync_plan_position();
+              stepper.synchronize();
+              hotend_duplication_enabled = true;
+              active_extruder_parked = false;
+              #if ENABLED(DEBUG_LEVELING_FEATURE)
+                if (DEBUGGING(LEVELING)) SERIAL_EM("Set hotend_duplication_enabled\nClear active_extruder_parked");
+              #endif
+            }
+            else {
+              #if ENABLED(DEBUG_LEVELING_FEATURE)
+                if (DEBUGGING(LEVELING)) SERIAL_EM("Active extruder not 0");
+              #endif
+            }
+            break;
+        }
+      }
+      return false;
+    }
+
+  #endif
 
   /**
    * line_to_current_position
@@ -680,403 +760,17 @@
     return position_is_reachable_raw_xy(RAW_X_POSITION(lx), RAW_Y_POSITION(ly));
   }
 
-  #if ENABLED(HYSTERESIS)
+  #if ENABLED(DUAL_X_CARRIAGE)
 
-    void Core_Mechanics::set_hysteresis(float x_mm, float y_mm, float z_mm, float e_mm) {
-      m_hysteresis_mm[X_AXIS] = x_mm;
-      m_hysteresis_mm[Y_AXIS] = y_mm;
-      m_hysteresis_mm[Z_AXIS] = z_mm;
-      m_hysteresis_mm[E_AXIS] = e_mm;
-
-      m_hysteresis_bits = ((m_hysteresis_mm[X_AXIS] != 0.0f) ? (1 << X_AXIS) : 0)
-                        | ((m_hysteresis_mm[Y_AXIS] != 0.0f) ? (1 << Y_AXIS) : 0)
-                        | ((m_hysteresis_mm[Z_AXIS] != 0.0f) ? (1 << Z_AXIS) : 0)
-                        | ((m_hysteresis_mm[E_AXIS] != 0.0f) ? (1 << E_AXIS) : 0);
-
-      calc_hysteresis_steps();
-    }
-
-    void Core_Mechanics::calc_hysteresis_steps() {
-      for (uint8_t i = 0; i < NUM_AXIS; i++)
-        m_hysteresis_steps[i] = (long)(m_hysteresis_mm[i] * planner.axis_steps_per_mm[i]);
-    }
-
-    void Core_Mechanics::set_hysteresis_axis(uint8_t axis, float mm) {
-      m_hysteresis_mm[axis] = mm;
-
-      if (mm != 0.0f) m_hysteresis_bits |=  ( 1 << axis);
-      else            m_hysteresis_bits &= ~( 1 << axis);
-
-      calc_hysteresis_steps();
-      report_hysteresis();
-    }
-
-    void Core_Mechanics::report_hysteresis() {
-      SERIAL_MV("Hysteresis X", m_hysteresis_mm[X_AXIS]);
-      SERIAL_MV(" Y", m_hysteresis_mm[Y_AXIS]);
-      SERIAL_MV(" Z", m_hysteresis_mm[Z_AXIS]);
-      SERIAL_MV(" E", m_hysteresis_mm[E_AXIS]);
-      SERIAL_MV(" SHIFTS: x=", m_hysteresis_axis_shift[X_AXIS]);
-      SERIAL_MV(" y=", m_hysteresis_axis_shift[Y_AXIS]);
-      SERIAL_MV(" z=", m_hysteresis_axis_shift[Z_AXIS]);
-      SERIAL_EMV(" e=", m_hysteresis_axis_shift[E_AXIS]);
-    }
-
-    void Core_Mechanics::insert_hysteresis_correction(const float x, const float y, const float z, const float e) {
-      long target[NUM_AXIS] = {x * planner.axis_steps_per_mm[X_AXIS], y * planner.axis_steps_per_mm[Y_AXIS], z * planner.axis_steps_per_mm[Z_AXIS], e * planner.axis_steps_per_mm[E_AXIS + active_extruder]};
-      uint8_t direction_bits = calc_direction_bits(planner.position, target);
-      uint8_t move_bits = calc_move_bits(planner.position, target);
-
-      // if the direction has changed in any of the axis that need hysteresis corrections...
-      uint8_t direction_change_bits = (direction_bits ^ m_hysteresis_prev_direction_bits) & move_bits;
-
-      if ((direction_change_bits & m_hysteresis_bits) != 0 ) {
-        // calculate the position to move to that will fix the hysteresis
-        for (uint8_t axis = 0; axis < NUM_AXIS; axis++) {
-          // if this axis changed direction...
-          if (direction_change_bits & (1 << axis)) {
-            long fix = (((direction_bits & (1 << axis)) != 0) ? -m_hysteresis_steps[axis] : m_hysteresis_steps[axis]);
-            //... add the hysteresis: move the current position in the opposite direction so that the next travel move is longer
-            planner.position[axis] -= fix;
-            m_hysteresis_axis_shift[axis] += fix;
-          }
-        }
-      }
-      m_hysteresis_prev_direction_bits = (direction_bits & move_bits) | (m_hysteresis_prev_direction_bits & ~move_bits);
-    }
-    
-    // direction 0: positive, 1: negative
-    uint8_t Core_Mechanics::calc_direction_bits(const long *position, const long *target) {
-      unsigned char direction_bits = 0;
-
-      if (target[X_AXIS] < position[X_AXIS])
-        direction_bits |= (1 << X_AXIS);
-      if (target[Y_AXIS] < position[Y_AXIS])
-        direction_bits |= (1 << Y_AXIS);
-      if (target[Z_AXIS] < position[Z_AXIS])
-        direction_bits |= (1 << Z_AXIS);
-      if (target[E_AXIS] < position[E_AXIS])
-        direction_bits |= (1 << E_AXIS);
-
-      return direction_bits;
-    }
-
-    uint8_t Core_Mechanics::calc_move_bits(const long *position, const long *target) {
-      uint8_t move_bits = 0;
-
-      if (target[X_AXIS] != position[X_AXIS])
-        move_bits |= (1 << X_AXIS);
-      if (target[Y_AXIS] != position[Y_AXIS])
-        move_bits |= (1 << Y_AXIS);
-      if (target[Z_AXIS] != position[Z_AXIS])
-        move_bits |= (1 << Z_AXIS);
-      if (target[E_AXIS] != position[E_AXIS])
-        move_bits |= (1 << E_AXIS);
-
-      return move_bits;
-    }
-
-  #endif
-
-  #if ENABLED(ZWOBBLE)
-
-    void Core_Mechanics::report_zwobble() {
-      if (!m_zwobble_sinusoidal)
-        SERIAL_M("Custom wobble function");
+    float Core_Mechanics::x_home_pos(const int extruder) {
+      if (extruder == 0)
+        return LOGICAL_X_POSITION(base_home_pos[X_AXIS]);
       else
-        SERIAL_MV("Core_Mechanics Amp(A): ", m_zwobble_amplitude);
-
-      SERIAL_MV(" phase(P): ", m_zwobble_phase); 
-      SERIAL_MV(" period(W): ", TWOPI / m_zwobble_puls);
-      SERIAL_MV(" puls: ", m_zwobble_puls);
-
-      if (!areParametersConsistent())
-        SERIAL_SM(WARNING, " Inconsistent parameters!");
-
-      SERIAL_E;
-
-      if (!m_zwobble_sinusoidal) {
-        // print out the LUT
-        for (int i = 0; i < zwobble_lutSize; i++) {
-          SERIAL_MV("Rod: ", ZROD(i));
-          SERIAL_MV(" Act: ", ZACTUAL(i));
-
-          int delta = (ZACTUAL(i) - ZROD(i)) * 200 + 20;
-          for (int j = 0; j < delta; j++) {
-            SERIAL_M(" ");
-          }
-          SERIAL_EM("  +");
-        }
-      }
-    }
-
-    // insert a planner.buffer_line if required to handle any hysteresis
-    void Core_Mechanics::insert_zwobble_correction(const float targetZ) {
-
-      if (!m_zwobble_consistent) return; // don't go through consistency checks all the time; just check one bool
-
-      float originZ = (float)planner.position[Z_AXIS] / planner.axis_steps_per_mm[Z_AXIS];
-
-      if (originZ < ZWOBBLE_MIN_Z || targetZ < ZWOBBLE_MIN_Z) return;
-
-      if (DEBUGGING(ALL)) {
-        SERIAL_MV("Origin: ", originZ);
-        SERIAL_MV(" Target: ", targetZ);
-      }
-
-      if (EQUAL_WITHIN_TOLERANCE(originZ, targetZ)) return; // if there is no Z move, do nothing
-
-      float originZRod;
-
-      // there is a high chance that the origin Z is the same as the last target Z: skip one iteration of the algorithm if possible
-      if (originZ == zwobble_lastZ)
-        originZRod = zwobble_lastZRod;
-      else
-        originZRod = findZRod(originZ);
-
-      if (DEBUGGING(ALL))
-        SERIAL_MV(" Origin rod: ", originZRod);
-
-      float targetZRod = findZRod(targetZ);
-
-      if (DEBUGGING(ALL))
-        SERIAL_MV(" Target Rod: ", targetZRod);
-
-      // difference in steps between the correct movement (originZRod->targetZRod) and the planned movement
-      long stepDiff = LROUND((targetZRod - originZRod) * planner.axis_steps_per_mm[Z_AXIS]) - (LROUND(targetZ * planner.axis_steps_per_mm[Z_AXIS]) - planner.position[Z_AXIS]);
-
-      if (DEBUGGING(ALL))
-        SERIAL_EMV(" stepDiff: ", stepDiff);
-
-      zwobble_lastZ = targetZ;
-      zwobble_lastZRod = targetZRod;
-
-      // don't adjust if target posizion is less than 0
-      if (planner.position[Z_AXIS] - stepDiff > 0)
-        planner.position[Z_AXIS] -= stepDiff;
-    }
-
-    void Core_Mechanics::set_zwobble_amplitude(float _amplitude) {
-      m_zwobble_amplitude = _amplitude;
-      m_zwobble_sinusoidal = true; // set_zwobble_amplitude sets to sinusoidal by default
-      calculateLut();
-    }
-
-    void Core_Mechanics::set_zwobble_period(float _period) {
-      if (_period <= 0) return;
-      m_zwobble_puls = TWOPI / _period;
-      calculateLut();
-    }
-
-    void Core_Mechanics::set_zwobble_phase(float _phase ) {
-      // poor man's modulo operation
-      while (_phase > 0) _phase -= 360;
-      while (_phase < 0) _phase += 360;
-
-      // phase now will be between 0 and 360
-      m_zwobble_phase = RADIANS(_phase); // convert phase to radians
-    }
-
-    void Core_Mechanics::set_zwobble_sample(float zRod, float zActual) {
-      if (DEBUGGING(ALL)) {
-        SERIAL_MV("New sample Rod: ", zRod);
-        SERIAL_EMV(" Act: ", zActual);
-      }
-
-      if (m_zwobble_puls <= 0) {
-        SERIAL_EM("You must define a period first (M97 W...)");
-        return;
-      }
-
-      if (m_zwobble_sinusoidal) {
-        m_zwobble_sinusoidal = false;
-        calculateLut(); // initializes the LUT to linear
-      }
-      insertInLut(zRod, zActual);
-    }
-
-    void Core_Mechanics::set_zwobble_scaledsample(float zRod, float zScaledLength) {
-      // We want to be able to correct scaling factor or set it before/after samples, so (ICK) store scaled samples as negative numbers  
-      set_zwobble_sample(zRod, -zScaledLength);
-
-      // Make sure we have a non-zero scaling factor
-      if (!m_zwobble_scalingFactor) {
-        m_zwobble_scalingFactor = 1.0;
-      }
-
-      // Find two scaled samples close to period
-      float period = TWOPI / m_zwobble_puls;
-      int s1 = -1, s2 = -1;
-
-      for (int i = 0; i < zwobble_lutSize; i++) {
-        if (ZACTUAL_IS_SCALED(i)) {
-          s1 = s2;
-          s2 = i;
-          if (ZROD(s2) >= period) break;
-        }
-      }
-
-      // Calculate scaling factor so zact[period] = zrod[period]
-      if (s2 >= 0 && ZACTUAL(s2)) {
-        // Case 1 - Only one sample
-        if (s1 < 0) {
-          m_zwobble_scalingFactor *= ZROD(s2) / ZACTUAL(s2);
-        }
-
-        // Case 2 - Samples bracketing period (s1 - p - s2): average zact
-        else if (ZROD(s2) > period) {
-          float gap1 = period - ZROD(s1);
-          float gap2 = ZROD(s2) - period;
-          float zActPeriod = ZACTUAL(s1) + (ZACTUAL(s2) - ZACTUAL(s1)) * gap1 / (gap1 + gap2);
-
-          m_zwobble_scalingFactor *= period / zActPeriod;
-        }
-
-        // Case 3 - Both samples before period (s1 - s2 - p): extrapolate zact
-        else {
-          float gap1 = ZROD(s2) - ZROD(s1);
-          float gap2 = period - ZROD(s2);
-          float zActPeriod = ZACTUAL(s2) + (ZACTUAL(s2) - ZACTUAL(s1)) * gap2 / gap1;
-
-          m_zwobble_scalingFactor *= period / zActPeriod;
-        }
-      }
-    }
-
-    void Core_Mechanics::set_zwobble_scalingfactor(float zActualPerScaledLength) {
-      m_zwobble_scalingFactor = zActualPerScaledLength;
-    }
-
-    // calculate the ZRod -> Zactual LUT using the model Zactual = Zrod + sin(w*Zrod) - this will actually only be used for one period
-    void Core_Mechanics::calculateLut() {
-      zwobble_lastZ = -1.0;
-      zwobble_lastZRod = -1.0; // reinitialize memorized Z values since we are changing the model
-      if (!areParametersConsistent()) return;
-      if (!m_zwobble_sinusoidal) {
-        initLinearLut();
-        return; // if the model is not sinusoidal, initializes LUT to linear
-      }
-      zwobble_lutSize = STEPS_IN_ZLUT;
-      float period = TWOPI / m_zwobble_puls;
-      // divide the period in STEPS_IN_ZLUT steps
-      float lutStep = period / STEPS_IN_ZLUT;
-      for (int i = 0; i < STEPS_IN_ZLUT; i++) {
-        float zRod = lutStep * i;
-        SET_ZROD(i, zRod);
-        SET_ZACTUAL(i, zRod + m_zwobble_amplitude * sin(m_zwobble_puls * zRod));
-      }
-    }
-
-    void Core_Mechanics::initLinearLut() {
-      float period = TWOPI / m_zwobble_puls;
-      zwobble_lutSize = 2; // only 2 samples originally
-      SET_ZROD(0, 0);
-      SET_ZACTUAL(0, 0);
-      SET_ZROD(1, period);
-      SET_ZACTUAL(1, period);
-    }
-
-    void Core_Mechanics::insertInLut(float zRod, float zActual) {
-      // check if the given zRod alread exists in LUT
-      for (int i = 0; i < zwobble_lutSize; i++) {
-        if (EQUAL_WITHIN_TOLERANCE(zRod, ZROD(i))) {
-          // replace value
-          SET_ZROD(i, zRod);
-          SET_ZACTUAL(i, zActual);
-          return;
-        }
-      }
-
-      // ok the value does not exist: is there still space in LUT? Insert it
-      if (zwobble_lutSize < STEPS_IN_ZLUT) {
-        int zPlace = -1;
-        for (int i = 0; i < zwobble_lutSize; i++) {
-          if (ZROD(i) > zRod) {
-            zPlace = i;
-            break;
-          }
-        }
-
-        // shift samples after zPlace
-        for (int i = zwobble_lutSize; i > zPlace; i--) {
-          SET_ZROD(i, ZROD(i - 1));
-          SET_ZACTUAL(i, ZACTUAL(i - 1));
-        }
-        zwobble_lutSize++; // increase zwobble_lutSize
-
-        // insert sample
-        SET_ZROD(zPlace, zRod);
-        SET_ZACTUAL(zPlace, zActual);
-        return;
-      }
-      else {
-        // zwobble_lutSize == STEPS_IN_ZLUT: replace the closest point with the new sample
-        int zPlace = 0;
-        float dist = DISTANCE(zRod, ZROD(zPlace));
-        for (int i = 1; i < zwobble_lutSize; i++) {
-          if (DISTANCE(zRod, ZROD(i)) < dist) {
-            zPlace = i;
-            dist = DISTANCE(zRod, ZROD(i));
-          }
-        }
-
-        SET_ZROD(zPlace, zRod);
-        SET_ZACTUAL(zPlace, zActual);
-      }
-    }
-
-    float Core_Mechanics::findInLut(float z) {
-      int i = 0;
-      if (z >= ZACTUAL(zwobble_lutSize - 1))
-        return ZROD(zwobble_lutSize - 1);
-
-      if (z <= ZACTUAL(0))
-        return ZROD(0);
-
-      for (i = 0; i < zwobble_lutSize; i++) {
-        if (ZACTUAL(i) > z)
-          break;
-      }
-
-      float invZDist = 1 / (ZACTUAL(i) - ZACTUAL(i-1)); // distance between Z steps
-
-      float interpZ = (ZROD(i - 1) * (ZACTUAL(i) - z) + ZROD(i) * (z - ZACTUAL(i - 1))) * invZDist; // linear interpolation between neighboring Z values
-
-      return interpZ;
-    }
-
-    // Find the Z value to be given to the "rod" in order to obtain the desired Z
-    float Core_Mechanics::findZRod(float z) {
-      int nCycle = 0;
-      float identicalZ = -m_zwobble_phase / m_zwobble_puls;
-
-      // find the last point in which the two Z are identical: this happens every (2kPI-phase)/w
-      while (identicalZ <= z) identicalZ = (TWOPI * (++nCycle) - m_zwobble_phase) / m_zwobble_puls; 
-
-      // find Z again using the previous cycle
-      identicalZ = (TWOPI *(nCycle - 1) - m_zwobble_phase) / m_zwobble_puls;
-
-      float deltaZa = z - identicalZ;
-
-      // find deltaZRod by linear interpolation of the lookup table
-      float deltaZrod = findInLut(deltaZa);
-      return identicalZ + deltaZrod;
-    }
-
-    bool Core_Mechanics::areParametersConsistent() {
-      if (!m_zwobble_sinusoidal) {
-        m_zwobble_consistent = true;  // parameters are always consistent if lut is not sinusoidal
-        return true;          // if the model is not sinusoidal, then don't check for consistency
-      }
-
-       // m_zwobble_amplitude*m_zwobble_puls must be less than 1 in order for the function to be invertible (otherwise it would mean that the wobble is so much that the axis goes back)
-      if (m_zwobble_puls <= 0 || m_zwobble_amplitude <= 0 || (m_zwobble_amplitude * m_zwobble_puls) >= 1) {
-        m_zwobble_consistent = false;
-        return false;
-      }
-
-      m_zwobble_consistent = true;
-      return true;
+        // In dual carriage mode the extruder offset provides an override of the
+        // second X-carriage offset when homed - otherwise X2_HOME_POS is used.
+        // This allow soft recalibration of the second extruder offset position without firmware reflash
+        // (through the M218 command).
+        return LOGICAL_X_POSITION(hotend_offset[X_AXIS][1] > 0 ? hotend_offset[X_AXIS][1] : X2_HOME_POS);
     }
 
   #endif